LayoutTests/media/modern-media-controls/pip-support/pip-support-enabled-expected.txt

Testing the PipSupport behavior.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.pipButton.enabled is true
PASS mediaController.controls.pipButton.enabled is false

PASS successfullyParsed is true

TEST COMPLETE
